feat: expose package configparse: move package
- ./internal/configparse -> ./configparse
